# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bot Quà Tặng AI Admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  password: string;
}

export interface LoginResult {
  token: string;
}

export interface BotStats {
  totalUsers: number;
  stock: number;
  claimed: number;
  banned: number;
  roundId: string;
}

export interface BotSettings {
  shopLink: string;
  shopUsername: string;
  supportUsername: string;
  cooldownHours: number;
  roundId: string;
}

export interface BotSettingsInput {
  shopLink?: string;
  shopUsername?: string;
  supportUsername?: string;
  cooldownHours?: number;
  roundId?: string;
}

export interface Account {
  email: string;
  password: string;
}

export interface AccountsInput {
  accounts: Account[];
}

export interface AddAccountsResult {
  added: number;
  total: number;
}

export interface BotUser {
  userId: string;
  username: string;
  firstName: string;
  startedAt: string;
  banned: boolean;
}

export interface LogEntry {
  time: string;
  action: string;
  user: string;
  admin?: string;
}

export interface ClaimRecord {
  userId: string;
  username: string;
  firstName: string;
  claimTime: string;
  accountEmail: string;
  roundId: string;
}

export interface BroadcastInput {
  message: string;
}

export interface NewRoundInput {
  roundId: string;
}

export interface ActionResult {
  ok: boolean;
  message: string;
}

export type GetBotLogsParams = {
limit?: number;
};