# lib/api-zod/src/generated/types/accountsInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bot Quà Tặng AI Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Account } from './account';

export interface AccountsInput {
  accounts: Account[];
}