# lib/api-zod/src/generated/types/botSettingsInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bot Quà Tặng AI Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface BotSettingsInput {
  shopLink?: string;
  shopUsername?: string;
  supportUsername?: string;
  cooldownHours?: number;
  roundId?: string;
}