# lib/api-zod/src/generated/types/claimRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bot Quà Tặng AI Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimRecord {
  userId: string;
  username: string;
  firstName: string;
  claimTime: string;
  accountEmail: string;
  roundId: string;
}