# lib/api-zod/src/generated/types/logEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bot Quà Tặng AI Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface LogEntry {
  time: string;
  action: string;
  user: string;
  admin?: string;
}